requirement.txt
Django==2.0.2
django-ckeditor==5.4.0
django-crispy-forms==1.7.2
django-js-asset==1.0.0
django-recaptcha==1.4.0
Pillow==5.0.0
pytz==2018.3
